# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost'
    description: Generated server url
paths:
...
  /api/telegram/stop:
    post:
      tags:
        - telegram-client-controller
      summary: Остановить Telegram-клиент для команды
      operationId: stopClient
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Клиент остановлен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Нет доступа к команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
  /api/telegram/start:
    post:
      tags:
        - telegram-client-controller
      summary: Запустить Telegram-клиент для команды
      operationId: startClient
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Клиент запущен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Нет доступа к команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
...
  /api/telegram/status:
    get:
      tags:
        - telegram-client-controller
      summary: Получить статус Telegram-клиента команды
      operationId: getClientStatus
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Статус получен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TelegramClientStatusResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Нет доступа к команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
  /api/telegram/qr:
    get:
      tags:
        - telegram-client-controller
      summary: Получить QR-ссылку авторизации Telegram-клиента
      operationId: getQrLink
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ссылка получена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Нет доступа к команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
  /api/telegram/clients:
    get:
      tags:
        - telegram-client-controller
      summary: 'Список всех команд, где активен Telegram-клиент'
      operationId: listClients
      responses:
        '200':
          description: Список ID команд
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    TelegramClientStatusResponse:
      type: object
      properties:
        teamId:
          type: string
          format: uuid
        state:
          type: string
        running:
          type: boolean
    LinkResponse:
      type: object
      properties:
        link:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT